+++ shapes/Sphere.hpp
@@ -15,7 +15,7 @@ class Sphere : public Shape{
     public:
         Sphere(Direction axis, Point center);
         void setCoords(float data[6]);
-        int findIntersectionWithLine(Direction d, Point o, double result[]);
+        int findIntersectionWithLine(Direction d, Point o, float result[]);
         Direction getNormalAtPoint(Point p);
         void calculateTangentsAtPoint(Point p, Direction& t1, Direction& t2);
